Assets e resources ajustados
