+++ hydra/assets/ibge_api.py
@@ -1,26 +1,12 @@
-import pandas as pd  # Add new imports to the top of `assets.py`
+import pandas as pd
 from dagster import (
     AssetExecutionContext,
-    AssetOut,
     MetadataValue,
-    Output,
     asset,
     get_dagster_logger,
-    load_assets_from_modules,
-    multi_asset,
-)  # import the `dagster` library
+)
 
-from .resources import (
-    IBGE_api,
-    GeosampaClient,
-    )
-from . import censo_assets as _censo_assets
-
-from .utils.io.files import read_json
-
-CAMADAS = read_json(file='camadas.json', path='./hydra')
-
-censo_assets = load_assets_from_modules([_censo_assets])
+from ..resources import IBGE_api
 
 @asset(
     io_manager_key="bronze_io_manager",
@@ -100,33 +86,3 @@ def municipios_silver(
     )
 
     return mun_df.to_parquet()  # return df and the I/O manager will save it
-
-
-# Função auxiliar para criar as definições de assets com valores padrão
-def default_geosampa_bronze(**kwargs) -> AssetOut:
-    default = dict(
-        group_name="geosampa_bronze",
-        io_manager_key="bronze_io_manager",
-        dagster_type=dict,
-        is_required=False
-    )
-    default.update(kwargs)
-    return AssetOut(**default)
-
-
-@multi_asset(
-    outs={out: default_geosampa_bronze() for out in CAMADAS},
-    can_subset=True
-)
-def camadas_geosampa(
-    context: AssetExecutionContext,
-    geosampa_client: GeosampaClient
-):
-    for nome_camada in context.selected_output_names:
-        camada = geosampa_client.get_feature(nome_camada)
-        yield Output(
-            camada,
-            output_name=nome_camada,
-            metadata={
-                'núm. features': len(camada['features'])
-            })
