+++ hydra/resources/__init__.py
@@ -1,3 +1,3 @@
 from .ibge_api import IBGE_api
-from .censo_resource import CensoResource
-from .geosampa_client import GeosampaClient
+from .censo import CensoResource
+from .geosampa import GeosampaClient
